requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-database-url==1.0.3
django-environ==0.11.2
djangorestframework==3.15.1
gunicorn==21.2.0
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
django-jazzmin==3.0.0
whitenoise==6.6.0
dj-database-url==2.1.0